+++ src/main/java/gift/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package gift.dao;
+package gift.repository;
 
 import gift.vo.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
