+++ src/main/java/gift/repository/ProductMapper.java
@@ -1,4 +1,4 @@
-package gift.dao;
+package gift.repository;
 
 import gift.vo.Product;
